+++ src/OpenLoco/src/Ui/Windows/AboutMusic.cpp
@@ -78,7 +78,7 @@ namespace OpenLoco::Ui::Windows::AboutMusic
     }
 
     // 0x0043BFBB
-    static void getScrollSize(Ui::Window&, uint32_t, uint16_t*, uint16_t* const scrollHeight)
+    static void getScrollSize(Ui::Window&, uint32_t, uint32_t*, uint32_t* const scrollHeight)
     {
         *scrollHeight = numSongs * (kRowHeight * 3 + 4);
     }
