+++ src/OpenLoco/src/Ui/Windows/PromptBrowseWindow.cpp
@@ -235,7 +235,7 @@ namespace OpenLoco::Ui::Windows::PromptBrowse
     }
 
     // 0x004464A1
-    static void getScrollSize(Ui::Window& window, [[maybe_unused]] uint32_t scrollIndex, [[maybe_unused]] uint16_t* scrollWidth, uint16_t* scrollHeight)
+    static void getScrollSize(Ui::Window& window, [[maybe_unused]] uint32_t scrollIndex, [[maybe_unused]] uint32_t* scrollWidth, uint32_t* scrollHeight)
     {
         *scrollHeight = window.rowHeight * static_cast<uint16_t>(_files.size());
     }
